requirements.txt
anthropic>=0.40.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13
APScheduler>=3.10.4
requests>=2.31.0
python-dateutil>=2.8.2
rich>=13.7.0
prompt_toolkit>=3.0.43
gTTS>=2.5.0
playsound>=1.3.0
